pyproject.toml
[project]
name = "Liquimager"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5,<6",
    "pillow>=10,<11",
    "numexpr>=2,<3"
]